Linting

+++ soccer/planning/planner/PathTargetPlanner.cpp
@@ -33,7 +33,9 @@ Trajectory PathTargetPlanner::plan(const PlanRequest& request) {
 
     // Debug drawing
     if (request.debug_drawer != nullptr) {
-        request.debug_drawer->drawCircle(goalPoint, drawRadius, drawColor,
+        request.debug_drawer->drawCircle(goalPoint,
+                                         static_cast<float>(drawRadius),
+                                         drawColor,
                                          drawLayer);
     }
